last

+++ src/infrastructure/database/db_repository.py
@@ -23,6 +23,8 @@ class PredictionsRepository(PredictionRepositoryInterface):
         return predictions
     def delete_prediction_db(self, prediction_id: int, owner_id: int):
         prediction = self.db.query(PredictionsTable).filter(PredictionsTable.id == prediction_id, PredictionsTable.owner_id == owner_id).first()
+        if prediction is None:
+            return None
         self.db.delete(prediction)
         self.db.commit()
         return {"status": "borrado"}
